DDF-UI-117 Follow on to fix location of some code for the checkbox perf fix (#186)
- Some of the code was orphaned in an old folder
